server mainloop: use LinkedBlockingQueue instead of ConcurrentLinkedQueue
ConcurrentLinkedQueue, having no waiting ability, lead to the use of a
homebrew spinlock with a sleep time of 100ms. This meant that if the fuzzer
couldn't supply the server with another request immediately, it would
incur a 100ms wait, which would have the effect of throttling the fuzzing
throughput to 10/s.